refactor: split index.ts tool dispatcher by category
Pure mechanical refactor with zero behavior change. The 5,479-line
src/index.ts grew organically around a monolithic ListTools array + CallTool
switch statement that spanned ~3,600 lines. This commit carves those two
surfaces into thirteen per-category modules under src/tools/ while keeping
every pre/post-dispatch gate (account routing, agent-grant check,
permission preset, destructive-confirm, response-size guard, audit log)
in place inside the CallTool handler.

Structure:
  src/tools/types.ts        — ToolDef, ToolCallContext, ToolHandler, ToolModule
  src/tools/registry.ts     — concatenates defs in historical order + dispatch map
  src/tools/sending.ts      — send_email, reply_to_email, forward_email, send_test_email
  src/tools/reading.ts      — get_emails … fts_status, extract_meeting (early + late groups)
  src/tools/folders.ts      — get_folders, sync_folders, create_folder, delete_folder, rename_folder
  src/tools/actions.ts      — mark/star/move/bulk + label actions
  src/tools/deletion.ts     — delete_email, bulk_delete, bulk_delete_emails
  src/tools/analytics.ts    — get_email_stats, get_email_analytics, get_contacts, get_volume_trends
  src/tools/system.ts       — get_connection_status, sync_emails, clear_cache, get_logs
  src/tools/bridge.ts       — start_bridge, shutdown_server, restart_server
  src/tools/aliases.ts      — SimpleLogin alias tools
  src/tools/pass.ts         — Proton Pass pass_list, pass_search, pass_get
  src/tools/drafts.ts       — save_draft, schedule_email, reminders, cancel_scheduled_email
  src/tools/escalation.ts   — ALWAYS_AVAILABLE meta-tools (pre-gate)

Invariants preserved:
  • ListTools ordering: sending → reading-early → folders → actions →
    deletion → analytics → system → bridge → aliases → pass → drafts →
    reading-late → escalation (matches the pre-refactor file order so
    client-side system prompts built from the tool list do not change)
  • Wire protocol: every tool name, input schema, output schema,
    annotations object, and error response body is identical
  • Gate chain lives in index.ts — the per-category handlers only run
    after account routing + permission + destructive-confirm have passed
  • Module-level mutable state (bridgeAutoStarted, smtpStatus,
    analyticsCache, analyticsCacheInflight) unified into a single
    sharedState object so handler ctx and non-handler code see one source
    of truth

Verification:
  • npm run build → clean
  • npm test → 1,567 tests pass (unchanged)
  • npm audit → 0 vulnerabilities
  • Registry.allToolDefs().map(d => d.name) byte-identical to the pre-
    refactor name list

Line counts:
  src/index.ts: 5479 → 2055 (-3424)
  src/tools/*.ts: 3839 total across 14 new files
